Move abspath() to file_utils
